Providing filtering and ordering options.
Implementing record-based filtering (filtering by BrandId), field-based filtering (Choosing which fields to get in the response), as well as ordering. The implemented ordering allows the usesr to type the field that will be used for sorting the entities Ascending. An optional (-) symbol can be added before the field name to make the ordering Descending.